+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## v0.6.0-beta
+* Added support for Security Transfer transactions
+* Added extra validation for unsupported data
+
 ## v0.5.0-beta
 * Improved displaying of errors in case of parsing or calucation errors.
 * Adopted more accurate type definitions in the codebase.